move pokemon models to model project
