--- TPP.Common/PkmnModels/Gender.cs
@@ -1,11 +1,0 @@
-using System.Runtime.Serialization;
-
-namespace TPP.Common.PkmnModels
-{
-    [DataContract]
-    public enum Gender
-    {
-        [EnumMember(Value = "m")] Male,
-        [EnumMember(Value = "f")] Female,
-    }
-}
