--- TPP.Common/PkmnModels/README.md
@@ -1,5 +1,0 @@
-This directory contains POCOs for generic pokemon related data that adhere to the
- [pokecat specification](https://github.com/TwitchPlaysPokemon/pokecat/blob/master/unified_objects.md)
-and are therefore also suitable for exchange between systems.
-This includes proper serialization, if the serializer used recognizes data contract annotations
-(e.g. JSON.net does by default).
