+++ TPP.Core/Overlay/Events/MatchSettingUpEvent.cs
@@ -3,7 +3,6 @@ using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Runtime.Serialization;
 using Newtonsoft.Json;
-using TPP.Common.PkmnModels;
 using TPP.Model;
 
 namespace TPP.Core.Overlay.Events
